% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_bayes.R
\name{internal_bayes1arm_dpara}
\alias{internal_bayes1arm_dpara}
\title{Default design parameter for single arm}
\usage{
internal_bayes1arm_dpara()
}
\description{
Default design parameter for single arm
}
