% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_nb.R
\name{rcurrent_day_eos_1}
\alias{rcurrent_day_eos_1}
\title{Generate analysis data set}
\usage{
rcurrent_day_eos_1(data_full, min_fu_days = 12 * 7, pt_proportion = 1)
}
\description{
Generate analysis dataset for minimum fu days
}
