% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_rmap.R
\name{rmap_control_bin}
\alias{rmap_control_bin}
\title{Parametric approximation of robust meta-analytic predictive prior for multiple historical control data}
\usage{
rmap_control_bin(
  y,
  N,
  v = 1,
  m = 0,
  s = 1,
  weight_noninfo = 0.5,
  prior1_noninfo,
  prior2_noninfo
)
}
\description{
Parametric approximation of robust meta-analytic predictive prior for multiple historical control data
}
