% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_rmap.R
\name{rmap_diff}
\alias{rmap_diff}
\title{Parametric approximation of robust meta-analytic predictive prior for multiple historical differences between treatment and control}
\usage{
rmap_diff(
  y.dif,
  y.dif.se,
  v = 1,
  m = 0,
  s = 1,
  weight_noninfo = 0.5,
  prior1_noninfo,
  prior2_noninfo
)
}
\description{
Parametric approximation of robust meta-analytic predictive prior for multiple historical differences between treatment and control
}
