% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_class.R
\name{stb_create_design}
\alias{stb_create_design}
\title{Create a Study Design}
\usage{
stb_create_design(
  type = c("bayes_1arm", "bayes_2arm", "nb_borrow", "bayes_mstage_bin",
    "bayes_mstage_norm")
)
}
\description{
Create a Study Design
}
