% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_stan_shiny.R
\name{stb_shiny}
\alias{stb_shiny}
\title{Run Web-Based application}
\usage{
stb_shiny(appname = "mstage", pkgname = "simubayes2")
}
\description{
Call Shiny to run \code{statidea} as a web-based application.
}
\details{
A web browser will be brought up for users to access the GUI
}
