% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_stan_shiny.R
\name{stb_stan}
\alias{stb_stan}
\title{Call STAN models}
\usage{
stb_stan(
  lst_data,
  stan_mdl = c("nb_mix", "nb_pp", "b_pp", "n_pp"),
  chains = 4,
  iter = 2000,
  warmup = 1000,
  cores = 4,
  ...
)
}
\arguments{
\item{lst_data}{List of study data to be passed to STAN}

\item{stan_mdl}{STAN model name}

\item{chains}{STAN parameter. Number of Markov chainsm}

\item{iter}{STAN parameter. Number of iterations}

\item{warmup}{STAN parameter. Number of burnin.}

\item{...}{other options to call STAN sampling such as \code{thin},
\code{algorithm}. See \code{rstan::sampling} for details.#'}

\item{control}{STAN parameter. See \code{rstan::stan} for details.}
}
\value{
Result from STAN sampling
}
\description{
Call STAN models. Called by \code{psrwe_powerp}.
}
