% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_bayes.R, R/stb_tls_nb.R
\name{stb_tb_bayes_nb_borrow}
\alias{stb_tb_bayes_nb_borrow}
\title{Bayesian informative prior for negative binomial data}
\usage{
stb_tb_bayes_nb_borrow(
  cur_data,
  ext_prior,
  pri_mu0,
  pri_k,
  pri_log_rr_weak,
  weight,
  model = c("nb_mix", "nb_pp"),
  seed = NULL,
  ...
)

stb_tb_bayes_nb_borrow(
  cur_data,
  ext_prior,
  pri_mu0,
  pri_k,
  pri_log_rr_weak,
  weight,
  model = c("nb_mix", "nb_pp"),
  seed = NULL,
  ...
)
}
\arguments{
\item{pri_mu0}{}

\item{...}{Parameters for STAN sampling}
}
\description{
Draw posterior samples of log risk ratio for negative binomial study.

Draw posterior samples of log risk ratio for negative binomial study.
}
