% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_bayes.R
\name{stb_tl_bayes_beta_post}
\alias{stb_tl_bayes_beta_post}
\title{Get posterior distribution}
\usage{
stb_tl_bayes_beta_post(
  obs_y = 1,
  obs_n = 10,
  pri_ab = c(1, 1),
  x = seq(0, 1, by = 0.01),
  n_post = NULL
)
}
\description{
Get posterior distribution
}
