% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_pp.R
\name{stb_tl_bayes_bin_pp_control}
\alias{stb_tl_bayes_bin_pp_control}
\title{Parametric approximation of Bayesian power prior for multiple binomial historical control data}
\usage{
stb_tl_bayes_bin_pp_control(
  N,
  y,
  prior1,
  prior2,
  power.param,
  stan_mdl = "b_pp",
  ...
)
}
\description{
Parametric approximation of Bayesian power prior for multiple binomial historical control data
}
