% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_bayes.R
\name{stb_tl_bayes_decision}
\alias{stb_tl_bayes_decision}
\title{Bayesian decision making}
\usage{
stb_tl_bayes_decision(
  post_smp,
  decision_h0,
  decision_thresh,
  decision_gl = c("greater.than", "less.than")
)
}
\description{
Bayesian decision making
}
