% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_pp.R
\name{stb_tl_bayes_norm_pp_control}
\alias{stb_tl_bayes_norm_pp_control}
\title{Parametric approximation of Bayesian power prior for multiple normal historical control data}
\usage{
stb_tl_bayes_norm_pp_control(
  y.m,
  y.se,
  prior1,
  prior2,
  power.param,
  stan_mdl = "n_pp",
  ...
)
}
\description{
Parametric approximation of Bayesian power prior for multiple normal historical control data
}
