% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_pp.R
\name{stb_tl_bayes_para_pp}
\alias{stb_tl_bayes_para_pp}
\title{Parametric approximation of Bayesian power prior}
\usage{
stb_tl_bayes_para_pp(
  dtype = c("binary", "normal"),
  borrow = c("control", "diff", "logOR"),
  ...
)
}
\description{
Parametric approximation of Bayesian power prior
}
