% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_rmap.R
\name{stb_tl_bayes_para_rmap}
\alias{stb_tl_bayes_para_rmap}
\title{Parametric approximation of robust meta-analytic predictive prior}
\usage{
stb_tl_bayes_para_rmap(
  dtype = c("binary", "normal"),
  borrow = c("control", "diff", "logOR"),
  cores = 1,
  ...
)
}
\description{
Parametric approximation of robust meta-analytic predictive prior
}
