% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_pp.R
\name{stb_tl_bayes_pp_dif}
\alias{stb_tl_bayes_pp_dif}
\title{Parametric approximation of Bayesian power prior}
\usage{
stb_tl_bayes_pp_dif(
  y.dif,
  y.dif.se,
  prior1,
  prior2,
  power.param,
  stan_mdl = "n_pp",
  ...
)
}
\description{
Parametric approximation of Bayesian power prior for multiple historical
difference between treatment and control
}
