% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_nb.R
\name{stb_tl_rc_events_censor}
\alias{stb_tl_rc_events_censor}
\title{Censor Recurrent Event Data}
\usage{
stb_tl_rc_events_censor(dat_rc, t = 1)
}
\arguments{
\item{dat_rc}{recurrent events}

\item{t}{follow up time}
}
\description{
Censor Recurrent Event Data
}
