% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_nb.R
\name{stb_tl_rc_events_to_nb}
\alias{stb_tl_rc_events_to_nb}
\title{Convert Recurrent Events to NB}
\usage{
stb_tl_rc_events_to_nb(dat_rc, t = 1, var_id = "sid")
}
\arguments{
\item{dat_rc}{recurrent events data frame}

\item{t}{follow up time}
}
\description{
Convert Recurrent Events to NB
}
