% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_enroll.R
\name{stb_tl_simu_enroll}
\alias{stb_tl_simu_enroll}
\title{Simulate Enrollment Time}
\usage{
stb_tl_simu_enroll(
  n_pt_arm = 100,
  n_pt_tot = n_pt_arm * 2,
  par_enroll = list(type = "by_duration", pt_dur_mth = 24),
  mth_min_fu = NULL,
  mth_fix_fu = NULL,
  date_bos = "2022-01-01",
  mth_to_days = 30.4,
  ...
)
}
\arguments{
\item{mth_min_fu}{minimum follow-up months for all patients}

\item{mth_fix_fu}{fixed follow-up months for all patients}
}
\description{
Simulate Enrollment Time
}
