% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_enroll.R
\name{stb_tl_simu_enroll_arc}
\alias{stb_tl_simu_enroll_arc}
\title{Simulate Enrollment Time}
\usage{
stb_tl_simu_enroll_arc(
  ntot,
  enroll_dur_mth,
  min_fu = NULL,
  date_bos = NULL,
  mth_to_days = 30.4,
  ...
)
}
\description{
An archived version of enrollment simulation (archived function)
}
