% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_enroll.R
\name{stb_tl_simu_enroll_arms}
\alias{stb_tl_simu_enroll_arms}
\title{Simulate Enrollment Time}
\usage{
stb_tl_simu_enroll_arms(n_by_arm, ..., seed = NULL)
}
\description{
Simulate Enrollment Time for all Arms
}
