% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_enroll.R
\name{stb_tl_simu_enroll_by_center}
\alias{stb_tl_simu_enroll_by_center}
\title{Simulate Enrollment Time}
\usage{
stb_tl_simu_enroll_by_center(
  n_pt,
  n_pt_tot,
  n_center,
  pt_per_center_per_mth,
  center_per_mth
)
}
\description{
Simulate enrollment time by center
}
