% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_tls_enroll.R
\name{stb_tl_simu_enroll_by_dur}
\alias{stb_tl_simu_enroll_by_dur}
\title{Simulate Enrollment Time}
\usage{
stb_tl_simu_enroll_by_dur(n_pt, n_pt_tot, pt_dur_mth)
}
\description{
Simulate enrollment time by total time
}
