% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_toolkit.R
\name{tl_draw_arm_size}
\alias{tl_draw_arm_size}
\title{Draw sample size for each arm}
\usage{
tl_draw_arm_size(sample_size, ratio_by_arm, fixed = TRUE, seed = NULL)
}
\arguments{
\item{sample_size}{total sample size}

\item{ratio_by_arm}{randomization ratio by arm}

\item{fixed}{Fixed ratio or multinomial ratio}

\item{seed}{Random seed}
}
\description{
Draw sample size for each arm
}
